# docs/07_api/openapi.json
{
  "components": {
    "schemas": {
      "Approval": {
        "description": "One ``workflow.approvals`` row (data_model.md \u00a74.5).",
        "properties": {
          "approval_class": {
            "title": "Approval Class",
            "type": "string"
          },
          "approval_id": {
            "title": "Approval Id",
            "type": "string"
          },
          "context_digest": {
            "title": "Context Digest",
            "type": "string"
          },
          "decided_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Decided At"
          },
          "decided_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Decided By"
          },
          "decision_comment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Decision Comment"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "expires_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires At"
          },
          "options": {
            "items": {
              "type": "string"
            },
            "title": "Options",
            "type": "array"
          },
          "requested_at": {
            "format": "date-time",
            "title": "Requested At",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "step_id": {
            "title": "Step Id",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "workflow_id": {
            "title": "Workflow Id",
            "type": "string"
          }
        },
        "required": [
          "approval_id",
          "workflow_id",
          "step_id",
          "approval_class",
          "title",
          "description",
          "context_digest",
          "options",
          "status",
          "decided_by",
          "decision_comment",
          "requested_at",
          "expires_at",
          "decided_at"
        ],
        "title": "Approval",
        "type": "object"
      },
...
      "DecideApprovalRequest": {
        "description": "Mirrors :class:`~ai_os_kernel.workflow_engine.human_approval.\nDecision` \u2014 only ``approved``/``rejected`` are real decisions today\n(``human_approval.py``'s own disclosed scope: no ``changes_requested``\nhandling exists).",
        "properties": {
          "comment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Comment"
          },
          "decision": {
            "enum": [
              "approved",
              "rejected"
            ],
            "title": "Decision",
            "type": "string"
          }
        },
        "required": [
          "decision"
        ],
        "title": "DecideApprovalRequest",
        "type": "object"
      },
      "DecideApprovalResponse": {
        "description": "The real, recorded decision, plus whether \u2014 and how \u2014 this\nspecific workflow was genuinely resumed. ``resumed`` is ``False``\nwhenever the decided workflow is not ``se.delivery_pipeline`` (the\none real, wired resumption target today) \u2014 a real, honest signal,\nnot a placeholder.",
        "properties": {
          "approval_id": {
            "title": "Approval Id",
            "type": "string"
          },
          "decided_by": {
            "title": "Decided By",
            "type": "string"
          },
          "decision": {
            "title": "Decision",
            "type": "string"
          },
          "resumed": {
            "title": "Resumed",
            "type": "boolean"
          },
          "resumed_error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resumed Error"
          },
          "resumed_outcome": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/WorkflowRunOutcome"
              },
              {
                "type": "null"
              }
            ]
          },
          "workflow_id": {
            "title": "Workflow Id",
            "type": "string"
          }
        },
        "required": [
          "approval_id",
          "workflow_id",
          "decision",
          "decided_by",
          "resumed",
          "resumed_outcome",
          "resumed_error"
        ],
        "title": "DecideApprovalResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PendingApprovalsResponse": {
        "description": "api_architecture.md \u00a76.2's own documented ``GET /api/v1/approvals``\n(\"Pending approvals\") \u2014 the real gap this module's own sibling,\n``human_approval.py``, named as \"real, separate, later work\"\n(``P06-S03-M39-T02``). Deliberately unpaginated \u2014 see\n:meth:`~ai_os_kernel.workflow_engine.human_approval.\nSqlApprovalRepository.list_pending`'s own docstring for why the\npending queue does not need the cursor-paginated envelope\n``GET /workflows`` uses.",
        "properties": {
          "approvals": {
            "items": {
              "$ref": "#/components/schemas/Approval"
            },
            "title": "Approvals",
            "type": "array"
          }
        },
        "required": [
          "approvals"
        ],
        "title": "PendingApprovalsResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
      "WorkflowRunOutcome": {
        "description": "How a :meth:`WorkflowAdvanceRunner.run_to_completion` call ended.",
        "enum": [
          "completed",
          "max_iterations_reached",
          "failed",
          "waiting_for_human"
        ],
        "title": "WorkflowRunOutcome",
        "type": "string"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Transport layer for the AI_OS Platform Kernel. No business logic lives here \u2014 see docs/07_api/api_architecture.md.",
    "title": "AI_OS Kernel API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/approvals": {
      "get": {
        "operationId": "list_pending_approvals_api_v1_approvals_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PendingApprovalsResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Pending Approvals",
        "tags": [
          "approvals"
        ]
      }
    },
...
    "/api/v1/workflows/{workflow_id}/approvals/{approval_id}/decisions": {
      "post": {
        "operationId": "decide_approval_api_v1_workflows__workflow_id__approvals__approval_id__decisions_post",
        "parameters": [
          {
            "in": "path",
            "name": "workflow_id",
            "required": true,
            "schema": {
              "title": "Workflow Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "approval_id",
            "required": true,
            "schema": {
              "title": "Approval Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DecideApprovalRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DecideApprovalResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Decide Approval",
        "tags": [
          "approvals"
        ]
      }
    },
...
  }
}